Changes to controllers. Working better.

+++ backend/moviedb/src/main/java/com/group12/moviedb/services/UserService.java
@@ -23,7 +23,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(int id) {
         userRepository.deleteById(id);
     }
